Move code to index.js
…because we need access to the instance of the markdown-it parser in
order for the filter to work correctly. This was the least painful way
to get to that.
